Whitelist file extensions
Commonly, repos will contain files that are text-based but are not
source code, e.g. .svg, .md, or .txt. In most cases, these files will
not be interesting in the context of finding hot spots within an
application. By providing a file extension whitelist, the results should
be more relevant.